feat: simplify runtime permission attack
