+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
             android:screenOrientation="portrait"
             android:exported="false" />
         <activity
-            android:name=".runtimepermissions.TapGameActivity"
+            android:name=".runtimepermissions.RuntimePermissionAttackActivity"
             android:screenOrientation="portrait"
             android:exported="false" />
         <activity
