+++ app/src/main/res/layout/activity_runtime_permissions.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="12dp"
         android:textAlignment="center"
-        android:text="@string/game_desc"/>
+        android:text="@string/runtime_permission_attack_description"/>
 
     <Button
         android:id="@+id/btn_start_game"
@@ -21,7 +21,7 @@
         android:layout_marginTop="12dp"
         android:layout_marginHorizontal="12dp"
         android:backgroundTint="?attr/colorPrimary"
-        android:text="@string/start_game"
+        android:text="@string/start"
         android:textSize="18sp"
         android:layout_gravity="center"/>
